Adding Spark Env entries that define the IP/Hostnames for the Driver and Executors.
The IP addresses that a given Spark Driver/Application
uses to bind by default might not be the right ones for a
given environment. This will cause connection and routing
issues between the Spark Driver and its Executors.

To mitigate this problem we are now exposing the following
environment variables that could be set:

    SPARK_LOCAL_IP
    SPARK_PUBLIC_DNS